Refactor Keycloak reconciler into its own struct
